extract common functions to registrysweepers.utils
